[REF] Parser: make Parser into a class
